Reflect temp. unit in pref.
